TNP-1 feat: tabelas db de entrada e saída e TNP-2 formulário e lógica de entrada>
